chore: rename `gcs-resumable-upload` -> `resumable-upload` (#1950)
